==> taxi_demand_predictor/notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data
from src import paths
import pandas as pd

rides_2021 = load_raw_data(2021)
rides_2022 = load_raw_data(2022)
rides = pd.concat([rides_2021, rides_2022])

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data.head(8)
ts_data.shape

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len= 24*28*1, # to just one week
    step_size=24
)

print(features.shape)
print(targets.shape)

print(features.columns)
#print(targets.columns)



tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR

tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
